fix: lock atoms on rings and boost rotation
